Update Bear and Big Brother.cpp

+++ 600/Bear and Big Brother.cpp
@@ -5,7 +5,7 @@ using namespace std;
 
 int main()
 {
-    int a,b;
+    ll a,b;
     cin>>a>>b;
     int year=0;
     while(a<=b)
